+++ src/node.h
@@ -2471,9 +2471,11 @@ public:
 class CodeCompNode : public WithRtti<Node, NodeKind::CodeComp> {
 public:
     enum Kind : unsigned int {
-        VAR,    // complete variable names
-        MEMBER, // complete members (field or method)
-        TYPE,   // complete type name (may be type alias)
+        VAR,            // complete variable names
+        MEMBER,         // complete members (field or method)
+        TYPE,           // complete type name (may be type alias)
+        CMD_OR_STMT,    // complete command or statement
+        CMD_OR_EXPR,    // complete command or expression
     };
 
 private: